requirements.txt
alembic==1.13.1
aspose-words==24.1.0
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
Deprecated==1.2.14
docx2pdf==0.1.8
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
greenlet==3.0.3
gunicorn==22.0.0
humanfriendly==10.0
idna==3.6
imageio==2.33.1
img2pdf==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
lazy_loader==0.3
llvmlite==0.41.1
lxml==5.1.0
Mako==1.3.0
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
onnxruntime==1.16.3
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
PDFNetPython3==9.5.0.post1
pikepdf==8.11.2
Pillow==9.5.0
platformdirs==4.1.0
plum-dispatch==1.7.4
pooch==1.8.0
protobuf==4.25.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyMatting==1.1.12
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
referencing==0.32.1
rembg==2.0.53
requests==2.31.0
rpds-py==0.16.2
scikit-image==0.22.0
scipy==1.11.4
Spire.Doc==12.1.0
SQLAlchemy==2.0.25
sympy==1.12
tifffile==2023.12.9
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.2
